refactor: Split `KalmanVertexUpdater` by dimension
Similar to `GainMatrixUpdater` this splits the `KalmanVertexUpdater`
into separate compilation units for dim 3 and dim 4 vertices.
