+++ Core/src/Vertexing/CMakeLists.txt
@@ -3,6 +3,8 @@ target_sources(
   PRIVATE
     AdaptiveGridTrackDensity.cpp
     KalmanVertexUpdater.cpp
+    KalmanVertexUpdaterImpl3.cpp
+    KalmanVertexUpdaterImpl4.cpp
     FsmwMode1dFinder.cpp
     VertexingError.cpp
     IterativeVertexFinder.cpp
